+++ R/data.R
@@ -6,6 +6,9 @@
 #' \describe{
 #'   \item{elevation}{Elevation above sea level, in meters}
 #' }
+#' @examples
+#' plot(dem, text_values = TRUE, breaks = "equal", col = terrain.colors(11))
+
 
 "dem"
 
@@ -17,10 +20,12 @@
 #' \describe{
 #'   \item{elevation}{Elevation above sea level, in meters}
 #' }
+#' @examples
+#' plot(carmel, breaks = "equal", col = terrain.colors(11))
 
 "carmel"
 
-#' Digital Elevation Model of Mount Carmel
+#' Digital Elevation Model of Golan Heights
 #'
 #' A \code{stars} object representing a Digital Elevation Model (DEM) Digital Elevation Model of part of the Golan Heights and Lake Kinneret, at 90m resolution
 #'
@@ -28,6 +33,8 @@
 #' \describe{
 #'   \item{elevation}{Elevation above sea level, in meters}
 #' }
+#' @examples
+#' plot(golan, breaks = "equal", col = terrain.colors(11))
 
 "golan"
 
@@ -39,6 +46,8 @@
 #' \describe{
 #'   \item{refl}{Reflectance, numeric value between 0 and 1}
 #' }
+#' @examples 
+#' plot(landsat, breaks = "equal")
 
 "landsat"